+++ src/Engine/save_load_system.rs
@@ -34,6 +34,17 @@ impl Plugin for SaveLoadPlugin{
 }
 
 
+fn load_saved_game(
+    //app_state: &mut ResMut<NextState<AppState>>,
+    game_state: &mut ResMut<NextState<GameState>>,
+){
+    //app_state.set(AppState::Game);
+    game_state.set(GameState::LoadGame);
+    //load_game(app_state, game_state);
+}
+
+
+
 #[derive(Resource)]
 pub struct ShouldSave {
     pub to_save: bool
