+++ src/game/menus/gameover.rs
@@ -9,7 +9,7 @@ use crate::{
 
 use super::{
     //components::OnScreenMenu, 
-    clean_menu, mainmenu::menu_camera, OnScreenMenu};
+    clean_menu, menu_camera, OnScreenMenu};
