+++ src/game/menus/menu_builder.rs
@@ -25,3 +25,5 @@ impl Menu {
         menu
     }
 }
+
+
