+++ src/game/menus/victory.rs
@@ -9,7 +9,7 @@ use crate::{
 
 use super::{
     components::OnScreenMenu, 
-    mainmenu::menu_camera, clean_menu};
+    menu_camera, clean_menu};
